backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
playwright>=1.40.0
google-generativeai>=0.3.0
supabase>=2.0.3
pydantic>=2.5.0
pydantic[email]>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.1
bcrypt>=4.1.0
PyJWT>=2.8.0
SQLAlchemy>=2.0.25
psycopg[binary]>=3.1.18
celery>=5.3.6
redis>=5.0.1
pytest>=7.4.0
beautifulsoup4>=4.12.0
# ee